+++ packages/spacecat-shared-data-access/src/v2/models/index.d.ts
@@ -10,6 +10,8 @@
  * governing permissions and limitations under the License.
  */
 
+import type { ValidationError } from '../index.d.ts';
+
 /**
  * Interface representing a base model for interacting with a data entity.
  */
@@ -21,10 +23,16 @@ export interface BaseModel {
   save(): Promise<this>;
 }
 
+export interface MultiStatusCreateResult<T> {
+  createdItems: T[],
+  errorItems: { item: object, error: ValidationError }[],
+}
+
 /**
  * Interface representing an Opportunity model, extending BaseModel.
  */
-export interface Opportunity extends BaseModel {
+export interface Opportunity extends BaseModel { /* eslint-disable no-use-before-define */
+  addSuggestions(suggestions: object[]): Promise<MultiStatusCreateResult<Suggestion>>;
   // eslint-disable-next-line no-use-before-define
   getSuggestions(): Promise<Suggestion[]>;
   getSiteId(): string;
@@ -73,7 +81,8 @@ export interface Suggestion extends BaseModel {
  */
 export interface BaseCollection<T extends BaseModel> {
   findById(id: string): Promise<T>;
-  create(data: object): Promise<T>;
+  create(item: object): Promise<T>;
+  createMany(items: object[]): Promise<MultiStatusCreateResult<T>>;
 }
 
 /**
@@ -90,6 +99,7 @@ export interface OpportunityCollection extends BaseCollection<Opportunity> {
 export interface SuggestionCollection extends BaseCollection<Suggestion> {
   allByOpportunityId(opportunityId: string): Promise<Suggestion[]>;
   allByOpportunityIdAndStatus(opportunityId: string, status: string): Promise<Suggestion[]>;
+  bulkUpdateStatus(suggestions: Suggestion[], status: string): Promise<Suggestion[]>;
 }
 
 /**