+++ packages/spacecat-shared-data-access/src/v2/models/opportunity.model.js
@@ -23,6 +23,25 @@ import BaseModel from './base.model.js';
 
 class Opportunity extends BaseModel {
   /**
+   * Adds the given suggestions to this Opportunity. Sets this opportunity as the parent
+   * of each suggestion, as such the opportunity ID does not need to be provided.
+   *
+   * @async
+   * @param {Array<Object>} suggestions - An array of suggestion objects to add.
+   * @return {Promise<{ createdItems: BaseModel[],
+   * errorItems: { item: Object, error: ValidationError }[] }>} - A promise that
+   * resolves to an object containing the created suggestion items and any
+   * errors that occurred.
+   */
+  async addSuggestions(suggestions) {
+    const childSuggestions = suggestions.map((suggestion) => ({
+      ...suggestion,
+      [this.idName]: this.getId(),
+    }));
+    return this._getAssociation('SuggestionCollection', 'createMany', childSuggestions);
+  }
+
+  /**
    * Retrieves all Suggestion entities associated to this Opportunity.
    * @async
    * @returns {Promise<Array<Suggestion>>} - A promise that resolves to an array of Suggestion