+++ packages/spacecat-shared-data-access/src/v2/models/suggestion.collection.js
@@ -38,7 +38,7 @@ class SuggestionCollection extends BaseCollection {
    * Retrieves all Suggestion entities by their associated Opportunity ID.
    * @async
    * @param {string} opportunityId - The unique identifier of the associated Opportunity.
-   * @returns {Promise<Array<Suggestion>>} - A promise that resolves to an array of Suggestion
+   * @returns {Promise<Suggestion[]>} - A promise that resolves to an array of Suggestion
    * instances related to the given Opportunity ID.
    * @throws {Error} - Throws an error if the opportunityId is not provided or if the query fails.
    */
@@ -74,6 +74,37 @@ class SuggestionCollection extends BaseCollection {
     ).go();
 
     return this._createInstances(records);
+  }
+
+  /**
+   * Updates the status of multiple given suggestions. The given status must conform
+   * to the status enum defined in the Suggestion schema.
+   * Saves the updated suggestions to the database automatically.
+   * You don't need to call save() on the suggestions after calling this method.
+   * @async
+   * @param {Suggestion[]} suggestions - An array of Suggestion instances to update.
+   * @param {string} status - The new status to set for the suggestions.
+   * @return {Promise<*>} - A promise that resolves to the updated suggestions.
+   * @throws {Error} - Throws an error if the suggestions are not provided
+   * or if the status is invalid.
+   */
+  async bulkUpdateStatus(suggestions, status) {
+    if (!Array.isArray(suggestions)) {
+      throw new Error('Suggestions must be an array');
+    }
+
+    const validStatuses = this._getEnumValues('status');
+    if (!validStatuses?.includes(status)) {
+      throw new Error(`Invalid status: ${status}. Must be one of: ${validStatuses.join(', ')}`);
+    }
+
+    suggestions.forEach((suggestion) => {
+      suggestion.setStatus(status);
+    });
+
+    await this._saveMany(suggestions);
+
+    return suggestions;
   }
 }